requirements.txt
streamlit
yfinance
pandas
numpy
plotly
scikit-learn
vaderSentiment
openai
requests
python-dotenv
